Began work on converting int size

+++ src/main.h
@@ -7,14 +7,14 @@
 typedef uint32_t vec_t;
 
 void MergePathSplitter(
-    vec_t * A, uint32_t A_length,
-    vec_t * B, uint32_t B_length,
-    vec_t * C, uint32_t C_length,
+    vec_t * A, uint64_t A_length,
+    vec_t * B, uint64_t B_length,
+    vec_t * C, uint64_t C_length,
     uint32_t threads, uint32_t* ASplitters, uint32_t* BSplitters);
     void MergePathSplitter2(
-        vec_t * A, uint32_t A_length,
-        vec_t * B, uint32_t B_length,
-        vec_t * C, uint32_t C_length,
+        vec_t * A, uint64_t A_length,
+        vec_t * B, uint64_t B_length,
+        vec_t * C, uint64_t C_length,
         uint32_t threads, uint32_t* ASplitters, uint32_t* BSplitters, uint32_t p);
 
 int hostBasicCompare(const void * a, const void * b);
